unify find feature model
